libmain/progress-bar: implement synchronized updates
If tearing occurs while the progress bar is rendering, flickering can occur.
This is particularly visible in multiline mode. Use the synchronized updates
specification [1] to make updates atomic on supported terminal emulators. On
unsupported terminals, the escape sequences should be ignored, leaving the
behaviour effectively unchanged (but there's nothing we could do better in this
situation).

[1] https://gitlab.com/gnachman/iterm2/-/wikis/synchronized-updates-spec

Change-Id: Ic91ca40422c73ac9f9e088c7cb78c1b8c28adc4e