manual: link directly to gerrit for proposing changes
This is, well, interesting. It's the same "edit" button you see on
Gitiles, and upon clicking it, without confirmation, it makes a new CL
and dumps you directly into the editor for the file in question.

I guess if you click it by accident you can just abandon or delete the
CL so it's like, fine. A bit of fascinating UX design here.

Change-Id: I765a1f1fc6d94b171878dde7b01fc621adc00180