main: avoid crashing when aborting completions
To print completions lix created a Finally object containing the actual function,
so the function was executed by the destructor of the class.
Unfortunately  aborting autocomplete by sending a SIGINT signal
(i.E. by pressing C-c) leads to an exception, that finally cant return or eat,
when throwing its own exception.

To avoid crashing when using auto complete let the function "mainWrapped" execute
the autocomplete code directly before returning.
This avoids creating the "Finally" object and instead moves the codeblock next to
the check to return when "arg.completions" is called.

Change-Id: Id333a60ad43c6095e8866f6953af78d51fd43b64