tests/functional2: add runner to justfile
Due to how meson works with the current justfile options, it is not
possible to pass additional arguments into the functional2 test
suit/pytest.
Due to that, it isn't possilbe to narrow down what tests to execute or
add output options or similar.

This commit adds an additional recipe, calling pytest directly ensuring
arguments are handed through

Change-Id: I3748d1cd5fddc16b11fff11c0f1a77195e37c837